payment_requirements.txt
flask==2.3.2
flask-cors==4.0.0
squareup==36.1.0.20240619
